moved store/lib to lib directory
