+++ core/lib/storage-manager.ts
@@ -1,6 +1,6 @@
 import { Logger } from '@vue-storefront/core/lib/logger'
 import * as localForage from 'localforage'
-import UniversalStorage from '@vue-storefront/core/store/lib/storage'
+import UniversalStorage from '@vue-storefront/core/lib/store/storage'
 import { currentStoreView } from '@vue-storefront/core/lib/multistore'
 import config from 'config'
