+++ core/lib/sync/task.ts
@@ -11,8 +11,8 @@ import EventBus from '@vue-storefront/core/compatibility/plugins/event-bus'
 import Task from '@vue-storefront/core/lib/sync/types/Task'
 import { Logger } from '@vue-storefront/core/lib/logger'
 import { TaskQueue } from '@vue-storefront/core/lib/sync'
-import * as entities from '@vue-storefront/core/store/lib/entities'
-import UniversalStorage from '@vue-storefront/core/store/lib/storage'
+import * as entities from '@vue-storefront/core/lib/store/entities'
+import UniversalStorage from '@vue-storefront/core/lib/store/storage'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
 import { processURLAddress } from '@vue-storefront/core/helpers'
 import { serial } from '@vue-storefront/core/helpers'
