+++ core/modules/catalog/helpers/search.ts
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import { Logger } from '@vue-storefront/core/lib/logger';
 import config from 'config';
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager';
-import { entityKeyName } from '@vue-storefront/core/store/lib/entities';
+import { entityKeyName } from '@vue-storefront/core/lib/store/entities';
 
 export const canCache = ({ includeFields, excludeFields }) => {
   const isCacheable = includeFields === null && excludeFields === null;
