+++ core/modules/catalog/store/category/mutations.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import { MutationTree } from 'vuex'
 import * as types from './mutation-types'
 import { slugify, formatBreadCrumbRoutes } from '@vue-storefront/core/helpers'
-import { entityKeyName } from '@vue-storefront/core/store/lib/entities'
+import { entityKeyName } from '@vue-storefront/core/lib/store/entities'
 import CategoryState from '../../types/CategoryState'
 import config from 'config'
 import { Logger } from '@vue-storefront/core/lib/logger'
