+++ core/modules/catalog/store/product/actions.ts
@@ -15,7 +15,7 @@ import { configureProductAsync,
   attributeImages } from '../../helpers'
 import { preConfigureProduct, getOptimizedFields, configureChildren, storeProductToCache, canCache, isGroupedOrBundle } from '@vue-storefront/core/modules/catalog/helpers/search'
 import SearchQuery from '@vue-storefront/core/lib/search/searchQuery'
-import { entityKeyName } from '@vue-storefront/core/store/lib/entities'
+import { entityKeyName } from '@vue-storefront/core/lib/store/entities'
 import { optionLabel } from '../../helpers/optionLabel'
 import { isOnline } from '@vue-storefront/core/lib/search'
 import omit from 'lodash-es/omit'
