+++ core/modules/catalog/store/tax/mutations.ts
@@ -1,6 +1,6 @@
 import { MutationTree } from 'vuex'
 import * as types from './mutation-types'
-import { entityKeyName } from '@vue-storefront/core/store/lib/entities'
+import { entityKeyName } from '@vue-storefront/core/lib/store/entities'
 import TaxState from '../../types/TaxState'
 import { Logger } from '@vue-storefront/core/lib/logger'
 import { StorageManager } from '@vue-storefront/core/lib/storage-manager'
