+++ core/modules/order/store/mutations.ts
@@ -1,6 +1,6 @@
 import { MutationTree } from 'vuex'
 import * as types from './mutation-types'
-import * as entities from '@vue-storefront/core/store/lib/entities'
+import * as entities from '@vue-storefront/core/lib/store/entities'
 import OrderState from '../types/OrderState'
 import config from 'config'
 import { Logger } from '@vue-storefront/core/lib/logger'
